requirements.txt
eventlet==0.33.3
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
WTForms==3.1.1
Flask-SocketIO==5.3.5
Flask-Limiter==3.5.0
Werkzeug==3.0.1
psycopg2-binary==2.9.9
python-socketio==5.10.0
python-engineio==4.8.0
gunicorn==21.2.0
Pillow==10.1.0
pywebpush==1.14.0
